BREAKS `error-data`: <key>* -> :err-<key> naming
This is sufficient for not conflicting with ex-info data in most
cases, and seems preferable to other alternatives like:

1. Namespaced keywords (:err/type, :err/message, :err/cause).
2. Special names (:type__, :msg__, :cause__), etc.
3. A nested map ({:type _ :msg _ :cause _ :data _}).

Which all offer reduced terseness / increased clumsiness at the
cost of _guaranteeing_ a non-conflicting merge.

Have also rearranged the data merge to make sure that any clash
that does happen results in ex-info's data winning.