Further memo gc optimization (against contention)
Previous atomic gc meant no possibility to gc a cache entry that
isn't an appropriate gc target (expired, etc.). But it also meant
high write-contention during GC (especially for ttl+max-n case
where even reads involve writes for tick counters, etc.).

This new approach offers *much* better performance in the face of
contention (the common case), at the cost of possibly gc'ing keys
that were an appropriate gc target at time of snapshot, but not
after. This should be rare, and not a big big deal when it happens.

An absolutely happy tradeoff.

Further work in this area may be to do actual `swap!` ops in key
batches to minimize the possibility of contention during bulk gc
key cleanup.