Invariants: remove auto `get-env` call
Even if we restrict this to the `:data` case, we'll be incurring a small
but significant cost even in the common case where we _don't_ want the
env map.

So the present solution is simple: just require an explicit call to
`(get-env)` in those cases when we want it. A little more verbose, but
means zero perf overhead in the common case.