Stubs: fix edgecase with same stub name in both clj+cljs
This was quite an interesting problem.

It effects .cljx and .cljc when there's a multi-platform stub defined
in the same namespace with the same name. The problem arises because
ClojureScript can have a macro and fn with the same name.

Specifically:

  - Say foo.cljc has `(defstub foostub)`
    - So foo.clj and foo.cljs *both* have a stub and unstub defined in the `foo` ns.
  - If consumer.cljs now wants to unstub foo and requires
    both foo.cljs and foo.clj (for macros), a call to `foo/unstub-foostub` is
    ambiguous. It could be a macro *or* fn call. ClojureScript will assume that the
    *macro* is being invoked.

The solution introduced in this PR is for the unstub *macro* to actually
support calling *both* from a Clojure *and* ClojureScript context.