Remove reflection warning caused by overloads introduced by Java 10
Java 10 added an overload with a Charset as second argument for these
methods:

 java.net.URLEncoder.encode(java.lang.String, java.lang.String)
 java.net.URLEncoder.encode(java.lang.String, java.nio.charset.Charset)

 java.net.URLDecoder.decode(java.lang.String, java.lang.String)
 java.net.URLDecoder.decode(java.lang.String, java.nio.charset.Charset)

My change removes the reflection warning by casting both arguments to
String in the #+clj case trusting that clojure prefers the String
variant.

Please note the asymmetry in the implementation of url-encode which
calls str on the first argument and url-decode which does not.