Fix: `cond!` confused by special keywords w/ odd number of clauses (@Knotschi)
Previously would have issues like the following:
  `(cond! :if-let [])` => `:else`

Also introduced a compile-time sanity check to ensure that there's a "then" clause
when using special keywords like `:if-let`, `:if-some`, `:if-none`.

Without this check, it was possible to produce technically correct but confusing
(and likely unintended) results like:
  `(cond :if-let [a :a] #_no-then-clause)` => nil

The above form will now throw a compile-time exception, and demand an explicit
"then" clause.